+++ openhands/server/routes/settings.py
@@ -69,140 +69,6 @@ async def load_settings(
         )
 
 
-@app.get('/secrets', response_model=GETSettingsCustomSecrets)
-async def load_custom_secrets_names(
-    settings: Settings | None = Depends(get_user_settings),
-) -> GETSettingsCustomSecrets | JSONResponse:
-    try:
-        if not settings:
-            return JSONResponse(
-                status_code=status.HTTP_404_NOT_FOUND,
-                content={'error': 'Settings not found'},
-            )
-
-        custom_secrets = []
-        if settings.secrets_store.custom_secrets:
-            for secret_name, _ in settings.secrets_store.custom_secrets.items():
-                custom_secrets.append(secret_name)
-
-        secret_names = GETSettingsCustomSecrets(custom_secrets=custom_secrets)
-        return secret_names
-
-    except Exception as e:
-        logger.warning(f'Invalid token: {e}')
-        return JSONResponse(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            content={'error': 'Invalid token'},
-        )
-
-
-@app.post('/secrets', response_model=dict[str, str])
-async def add_custom_secret(
-    incoming_secrets: POSTSettingsCustomSecrets,
-    settings_store: SettingsStore = Depends(get_user_settings_store),
-) -> JSONResponse:
-    try:
-        existing_settings = await settings_store.load()
-        if existing_settings:
-            for (
-                secret_name,
-                secret_value,
-            ) in existing_settings.secrets_store.custom_secrets.items():
-                if (
-                    secret_name not in incoming_secrets.custom_secrets
-                ):  # Allow incoming values to override existing ones
-                    incoming_secrets.custom_secrets[secret_name] = secret_value
-
-            # Create a new SecretStore that preserves provider tokens
-            updated_secret_store = UserSecrets(
-                custom_secrets=incoming_secrets.custom_secrets,
-                provider_tokens=existing_settings.secrets_store.provider_tokens,
-            )
-
-            # Only update SecretStore in Settings
-            updated_settings = existing_settings.model_copy(
-                update={'secrets_store': updated_secret_store}
-            )
-
-            await settings_store.store(updated_settings)
-
-        return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content={'message': 'Settings stored'},
-        )
-    except Exception as e:
-        logger.warning(f'Something went wrong storing settings: {e}')
-        return JSONResponse(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            content={'error': 'Something went wrong storing settings'},
-        )
-
-
-@app.delete('/secrets/{secret_id}')
-async def delete_custom_secret(
-    secret_id: str,
-    settings_store: SettingsStore = Depends(get_user_settings_store),
-) -> JSONResponse:
-    try:
-        existing_settings: Settings | None = await settings_store.load()
-        custom_secrets = {}
-        if existing_settings:
-            for (
-                secret_name,
-                secret_value,
-            ) in existing_settings.secrets_store.custom_secrets.items():
-                if secret_name != secret_id:
-                    custom_secrets[secret_name] = secret_value
-
-            # Create a new SecretStore that preserves provider tokens
-            updated_secret_store = UserSecrets(
-                custom_secrets=custom_secrets,
-                provider_tokens=existing_settings.secrets_store.provider_tokens,
-            )
-
-            updated_settings = existing_settings.model_copy(
-                update={'secrets_store': updated_secret_store}
-            )
-
-            await settings_store.store(updated_settings)
-
-        return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content={'message': 'Settings stored'},
-        )
-    except Exception as e:
-        logger.warning(f'Something went wrong storing settings: {e}')
-        return JSONResponse(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            content={'error': 'Something went wrong storing settings'},
-        )
-
-
-@app.post('/unset-settings-tokens', response_model=dict[str, str])
-async def unset_settings_tokens(
-    settings_store: SettingsStore = Depends(get_user_settings_store),
-) -> JSONResponse:
-    try:
-        existing_settings = await settings_store.load()
-        if existing_settings:
-            settings = existing_settings.model_copy(
-                update={'secrets_store': UserSecrets()}
-            )
-            await settings_store.store(settings)
-
-        return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content={'message': 'Settings stored'},
-        )
-
-    except Exception as e:
-        logger.warning(f'Something went wrong unsetting tokens: {e}')
-        return JSONResponse(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            content={'error': 'Something went wrong unsetting tokens'},
-        )
-
-
 @app.post('/reset-settings', response_model=dict[str, str])
 async def reset_settings() -> JSONResponse:
     """
